django/requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.1.15
django-allauth==0.42.0
django-cors-headers==3.3.0
django-extensions==2.2.9
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
Faker==4.1.0
gunicorn==20.0.4
idna==2.9
oauthlib==3.1.0
psycopg2==2.8.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
six==1.15.0
text-unidecode==1.3
urllib3==1.25.9
whitenoise==5.1.0
